+++ images/main/sv.sh
@@ -1,45 +1,23 @@
 #! /bin/bash
 
-
 APP="$1"
+#IMG="$U/$I:$APP"
 
 shift
 ARGS="$@"
 
 # USER_UID=$(id -u)
 
-#U=malex984
-#I=dockapp
-#IMG="$U/$I:$APP"
+# The following is an adaptation to the new naming schema: hilbert/$APP:$VERSION
+U=hilbert
+IMAGE_VERSION="${IMAGE_VERSION:-latest}"
+IMG="$U/${APP}:${IMAGE_VERSION}" # IMG="$APP" #IMG="$U/$I:$APP"
 
-ID=$(docker images | awk '{ print "[" $1 ":" $2 "]" }' | sort | uniq | grep "\[$APP\]" )
+ID=$(docker images | awk '{ print "[" $1 ":" $2 "]" }' | sort | uniq | grep "\[${IMG}\]")
 
-if [ ! -z "$ID" ]; then  
-  IMG="$APP"
-  APP=$(echo "$IMG" | sed 's@^.*:@@g')
-  if [ "$APP" = "latest" ]; then
-    APP=$(echo "$IMG" | sed -e 's@:.*$@@g' -e 's@/@_@g')  
-  fi
-else
-  ID=$(docker images | awk '{ print "[" $1 "]" }' | sort | uniq | grep "\[$APP\]" )
-  
-  if [ ! -z "$ID" ]; then  
-    IMG="$APP:latest"
-  else
- 
-    TAG=$(echo "$APP" | sed 's@^.*/.*:@@g')
-    
-    if [ "x$APP" = "x$TAG" ]; then 
-      # missing prefix for a missing standard image
-      IMG="$U/$I:$TAG"
-    else
-      # is it a full name for a missing image?
-      IMG="$APP"
-      APP="$TAG"
-    fi  
-    unset TAG
-  
-  fi  
+if [ -z "$ID" ]; then
+  echo "ERROR: no such image '${IMG}'"
+  exit 2
 fi
 
 APP="c_$APP"